+++ app/hooks/use-nprogress.ts
@@ -1,0 +1,28 @@
+import NProgress from "nprogress";
+import { useNavigation, useFetchers } from "@remix-run/react";
+import { useMemo, useEffect } from "react";
+
+export function useNprogress() {
+    let transition = useNavigation();
+
+    let fetchers = useFetchers();
+  
+    let state = useMemo<"idle" | "loading">(
+      function getGlobalState() {
+        let states = [
+          transition.state,
+          ...fetchers.map((fetcher) => fetcher.state),
+        ];
+        if (states.every((state) => state === "idle")) return "idle";
+        return "loading";
+      },
+      [transition.state, fetchers]
+    );
+  
+    useEffect(() => {
+      // waiting for the loaders of the next location so we start it
+      if (state === "loading") NProgress.start();
+      // when the state is idle then we can to complete the progress bar
+      if (state === "idle") NProgress.done();
+    }, [transition.state]);
+}
